requirements.txt
Django==4.0.3
djangorestframework==3.13.1
mysqlclient==2.1.0
psycopg2==2.9.3
django_cors_headers==3.11.0